+++ tests/providers/test_singleton.py
@@ -1,0 +1,25 @@
+import dataclasses
+
+from that_depends import BaseContainer, providers
+
+
+@dataclasses.dataclass(kw_only=True, slots=True)
+class SingletonFactory:
+    dep1: bool
+
+
+class DIContainer(BaseContainer):
+    singleton = providers.Singleton(SingletonFactory, dep1=True)
+
+
+async def test_singleton_provider() -> None:
+    singleton1 = await DIContainer.singleton()
+    singleton2 = await DIContainer.singleton()
+    singleton3 = DIContainer.singleton.sync_resolve()
+    await DIContainer.singleton.tear_down()
+    singleton4 = DIContainer.singleton.sync_resolve()
+
+    assert singleton1 is singleton2 is singleton3
+    assert singleton4 is not singleton1
+
+    await DIContainer.tear_down()
